+++ python/tvm/relay/build_module.py
@@ -26,11 +26,11 @@ from tvm.ir import IRModule
 from tvm.tir import expr as tvm_expr
 from .. import nd as _nd, target as _target, autotvm
 from ..contrib import graph_runtime as _graph_rt
-from ..runtime import graph_runtime_factory as _graph_runtime_factory
 from . import _build_module
 from . import ty as _ty
 from . import expr as _expr
 from . import function as _function
+from .backend import graph_runtime_factory as _graph_runtime_factory
 from .backend import interpreter as _interpreter
 from .backend.vm import VMExecutor
